ailment/hash: replace MD5 + byte-buffer with FxHasher
The previous ``stable_hash`` mirrored Python's
``angr.ailment.utils.stable_hash`` byte-for-byte and ran the result
through MD5 -- holding both an exact-format requirement and a
cryptographic-strength hash that the call site does not need.

Phase D collapsed the per-class Python AIL pyclasses into a single
Rust ``Expression`` / ``Statement`` pyclass, so there is no parallel
Python value whose hash needs to match: the Rust ``__hash__`` *is*
the AIL hash on rustyail. Anything that needed cross-language
compatibility has been gone for a while.

Rewrite to use ``rustc_hash::FxHasher`` (the hasher rustc itself uses
for symbol tables) directly via the std ``Hasher`` trait:

- ``hash.rs`` keeps the public ``HashItem`` / ``stable_hash`` API.
- Each ``HashItem`` variant emits a one-byte discriminant tag so
  ``Int(5)`` vs ``Bool(true)`` (or ``Str("Foo")`` vs ``TypeName("Foo")``)
  cannot collide via shared payload bytes.
- ``Str`` / ``TypeName`` / ``Raw`` use std's ``Hash`` impl for ``str``
  / ``&[u8]`` which writes a length prefix so concatenation cannot
  alias with a different split.
- ``Tuple`` is bracketed (open/close tag bytes) so a flat sequence
  cannot alias with a nested one.
- ``stable_hash`` now returns ``i64`` directly (was ``u32``; the
  previous ``as i64`` casts at every caller are dropped).

An intermediate xxh3-streaming variant was actually *slower* than
MD5 for typical AIL inputs (~2.4x): xxh3's streaming-state setup
overhead and the length-prefix padding on tiny payloads (~50 bytes
including tags vs MD5's single 64-byte block) outweighed the
algorithmic win. FxHasher sidesteps both: each ``write_u8`` is one
mul + one rotate + one xor with no streaming state, so the per-call
cost is roughly linear in input size with a very small constant.

Micro-bench (cold hash of a 3-node BinaryOp tree, per call):

- MD5 (before):              ~580 ns
- xxh3 streaming (rejected): ~1430 ns  (~2.4x slower than MD5)
- FxHasher (this commit):    ~172 ns  (~3.4x faster than MD5)

Touched in addition to ``hash.rs``:
- ``native/angr/Cargo.toml``: ``md-5`` removed; ``rustc-hash`` added.
- ``ail_expr.rs`` / ``ail_stmt.rs`` / ``block.rs``: trailing
  ``as i64`` casts dropped at the 42 ``stable_hash(...)`` call sites.

Unit tests in ``hash.rs`` cover the disambiguation properties
(distinct-variant-same-value, str-vs-typename, concat-vs-split,
nested-vs-flat, ``None`` stability). All pass.

Decompiler test suite (214/214) passes; doit walltime (N=20 medians):
master 1.30 s, rustyail pre-hash 1.60 s, this commit 1.58 s.

+++ native/angr/src/ailment/block.rs
@@ -266,7 +266,7 @@ impl Block {
             HashItem::TypeName("Block"),
             HashItem::Int(self.addr as i128),
             idx_item,
-        ]) as i64;
+        ]);
         self.cached_hash.set(h);
         // suppress the unused-import lint for Ordering when atomic isn't used directly
         let _ = Ordering::Relaxed;
